+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.3.0 - Unreleased
+## 0.3.0 - 2020-04-25
 ### Additions
 - Link to the GitHub repository from PyPI (#26). Thanks @theY4Kman!
 - Support for date properties (#30, #37). Thanks @acgray!